feat: Add character experience data

+++ Assets/_Project/Scripts/EnemySystem/EnemyTemplate.cs
@@ -40,6 +40,6 @@ namespace MightyAdventures.EnemySystem
     {
         public int level;
         public CharacterStats stats;
-        public int experience;
+        public float experience;
     }
 }
